+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.36.2
 * Changed output of `sugar.GenerateDeclareSection` (added error as second result)
 * Specified `sugar.GenerateDeclareSection` for `go1.18` (supports input types `*table.QueryParameters` `[]table.ParameterOption` or `[]sql.NamedArg`)
 * Supports different go's primitive value types as arg of `sql.Named("name", value)`